Move RepresentationType to its own file
RepresentationType is a type-level concept (resource type + optional
reporter), not tightly coupled to the instance-level ResourceReference.
Give it its own file per domain model conventions.

Made-with: Cursor
